requirements_dev.txt
pip==21.1
bump2version==1.0.1
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
tox==3.14.0
coverage==4.5.4
Sphinx==3.0.4
twine==2.0.0
httpx==0.18.1
python-dotenv==0.17.1
pydantic==1.8.2


pytest==4.6.5
pytest-runner==5.1
